# BackEnd/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": " EMS Management API",
    "version": "1.0.0",
    "title": "EMS Management API"
  },
  "host": "localhost:2228",
  "basePath": "/",
  "tags": [
    {
      "name": "login"
    },
    {
      "name": "plans"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "txt_usrId": {
                  "example": "any"
                },
                "txt_usr_pwd": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/signup/": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Learnings/importExcel": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/Learnings/dummyData": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/user/signUp": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pass": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/user/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "Pass": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/user/AddEmp/{UserId}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "UserId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "example": "any"
                },
                "Name": {
                  "example": "any"
                },
                "Address": {
                  "example": "any"
                },
                "MobileNo": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/SFTP/CallSFTP": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}